requirements.txt
Flask
SQLAlchemy
passlib[bcrypt]
python-dotenv
Werkzeug
psutil
Pillow
Jinja2
itsdangerous
click